backend/requirements.txt
# FastAPI and web framework
fastapi
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.12.1

# Caching and storage
redis==5.0.1
chromadb

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.1.4
chardet==5.2.0

# Machine learning and embeddings
sentence-transformers==2.2.2
huggingface-hub==0.16.4
numpy==1.24.4

# HTTP client for API calls
httpx==0.27

# Mistral AI client
mistralai

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
psutil==5.9.6
